linux-openxt: blktap2: Lock ring->vma mmap_sem not current's
Troy Crosley pointed out that the locking of current->mm->mmap_sem was
wrong since it was whatever thread was running at the time of the call
and not necessarily the tapdisk process.  The tapdisk's mm, store in
ring->vma->vm_mm is the one modified, so that is what needs locking.

WARN_ON(ring->vma->vm_mm != current->mm) in blktap_ring_map_segment()
triggered for openssl and kworker.

Push the locking down into blktap_ring_map_segment since that is the
only place that needs to modify the vma with vm_insert_page.  There is a
second call to vm_insert_page in blktap_ring_mmap(), but that one
should be called with mmap_sem already held.

Reported-by: Troy Crosley <crosleyt@ainfosec.com>
Signed-off-by: Jason Andryuk <jandryuk@gmail.com>