openxt distro conf: force exclude of OCaml xenstored to get C version
Setting both PREFERRED_PROVIDER_virtual/xenstored and
PREFERRED_RPROVIDER_virtual/xenstored is apparently not sufficient to
direct 'virtual/xenstored' to be resolved to the xen-tools C
implementation of xenstored, so force exclude the xen-ocaml-libs OCaml
version in the distro config where the preference is expressed.

Build tested and the rootfs inspected to verify that the C daemon is
installed when this is applied.

Signed-off-by: Christopher Clark <christopher.clark6@baesystems.com>