dom0: disable i915 PSR
Surfman with i915-foreign cannot support PSR. Since the framebuffer
pages are allocated in the guest memory space, but handled by Surfman in
dom0, this design does not allow frontend buffer tracking that would be
necessary for PSR.

In effect, this will boot to a black-screen or a grey screen with a
stuck pointer if a refresh happens on the first frames.

PSR requires both hardware and panel support, hence a not so obvious
list of platforms may reproduce the issue.

Disable i915 PSR in 5.x+ kernels (Surfman only supports foreign with
i915).

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>