layer/zeus: fix sub-project versions
Fix the revisions of sub-project to their stable version for this layer
stable branch.

In order to simplify version management, since multiple recipes can pull
a common repository, move the source/revision configuration to a
matching include file and sort the recipes per repository provider.

The current revisions are the tips of the master branches for each
subprojects.

Should the need for numbered version management per subproject arise, it
is now a matter of changing the relevant .inc file and submitting a PR
against this branch.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>