linux: blktap2: Fix poll wakeup and request flushing
flush_requests was added to the blktap2.patch, but it wasn't wired up to
actually trigger.  We should push the writes to the ring before we
notify userspace of new requests.  However, that won't work on its own
since poll would report no work and hang indefinitely.

The poll work determination was not idempotent since RING_PUSH_REQUESTS
updates req_prod, so future calls to poll would show work = 0.  Change
that to rsp_prod - req_prod since that will be true for outstanding
requests - ones that tapdisk can service.

Take the mutex lock when reading the rings.  It serialized access with
the production side.

xen ring.h has macros for handling this stuff, but tapdisk doesn't use
them or advance rsp_event or req_event.  The rsp/req_event is needed for
them to work on the driver side.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>