tpm-tools: Stop tpm-tools-sa depending on trousers
tpm-tools-sa takes pains to only link to libtspi-sa, but
tpm_unsealdata_sa also links to libtpm_unseal, and that linked to
libtspi and therefore pulled in all of trousers.

Create and statically link libtpm_unseal_sa which will use libtspi-sa
and avoid the dependency.

tpm-tools-quote-standalone.patch receives a fixup to apply cleanly.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>