tboot: Add `-C acm_caps` to pcr-calc
pcr-calc currently uses an ACM file to read out the SINIT capabilities
to determine its calculations.  Add an option to just directly provide
the SINIT capabilities.  This is useful for the case of a BIOS provided
ACM when a file is unavailable.  Otherwise pcr-calc will be unable to
run and you cannot forward seal.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>