measuredlaunch: Add txt-stat fallback for ACM caps
You can have a machine do a measured launch from a BIOS-provided ACM
when you don't have an SINIT ACM file.  In that case, you cannot forward
seal because ml-functions fails when it cannot find a file.

Add a fallback mode for when a file cannot be found to parse the SINIT
capabilities out of the txt-stat output.  It's not pretty, but it may do
the trick.  Then you can use the new -C option to pcr-calc for the
forward seal.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>