Revert "refpolicy: Silence mount AVC"
If mount is seeing device_t, something is wrong - Print the avc.

After the dunfell uprev, mount may have started failing from this
permission issue when trying to extract the kernel from a partitioned
disk.  This was seen with the usbvm in PV mode.  ndvm may also show it,
but it defaults to HVM.  An xenmgr change should fix this, so drop this
dontaudit.

This reverts commit a76ead77ddbe29abd40637d2aebeddee337e27e8.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>