busybox: Enable pkill for usbvm vusb-daemon
commit 5af2803853d6 "vusb: Fix initscript shutdown" switch to using
pkill for stopping vusb-daemon.  That is fine for dom0, but usbvm uses
busybox instead of coreutils, and the command wasn't build.  At shutdown
there would be a pkill command not found error.  Build pkill into
busybox so it works.

vusb-daemon shutdown didn't work before commit 5af2803853d6, so the lack
of pkill wasn't changing that.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>