layer: require out-of-tree signing key for modules
If module signing is enabled and key/cert are provided, add the
necessary dependencies between do_configure/do_sign_module to depend on
the out-of-tree key/cert.

If module signing is enabled and only cert is provided, add the
necessary dependency to do_configure, do_sign_module will not do
anything, offline signing is left to be done outside of OE.

If module signing is disabled, nothing is done, no dependency is added.

If module signing is enabled, but a certificate is not provided, error
out, this inevitably ends up re-generating a throw-away signing key when
out-of-tree modules dependencies are resolved (do_shared_workdir does
not store state, including signing keys).

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>