layer: use STAGING_KERNEL_BUILDDIR for defconfig
out-of-tree modules will not have .config in ${B}.
Check the defconfig in STAGING_KERNEL_BUILDDIR instead.

Add a bbnote, failing this test is not fatal if signing was disabled for
a kernel that inherited this class. More correctly this class should not
be inherited if the build kernel/modules are not intended to support
module signing.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>