rpc-autogen: module for haskell bindings
Provide generated Haskell bindings as a cabal module[1]
The bindings were generated and versioned using dbus-gen[2], then are
provided by the rpc-autogen recipe.
This makes it easier to track changes across versions and avoid
depending on generated sources in-tree for the haskell components.

The goal of this is to provide a reasonably easy to modify alternative
to rpcgen that could still with low efforts target different languages.

In this current implementation, only haskell is generated by dbus-gen,
the other generated bindings (for OCaml, C and JavaScript) still rely on
rpcgen.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>

[1] https://github.com/eric-ch/dbus-gen-bindings
[2] https://github.com/eric-ch/dbus-gen/tree/haskell-bindings