tpm-scripts: fix tpm_get_pcr grep format
For TPM1.2, /sys/class/tpm/tpm0/pcrs is formatted with:
'PCR-%02d: %02X %02X...'
The use of grep does not account for the 0 padding for PCRs below 10,
which breaks forward seal on UEFI/TPM1.2 systems.
The bug is mostly silent (apart from failing to unseal after
upgrading/forward-sealing):
-> tpm_get_pcr returns an empty string for PCR 8
-> ! contains_only "$pcr8" "0" is used to detect a UEFI measured
  installation. "contains_only" will return true for
  an empty-string given for the PCR value, for any character value.
-> In turn, seal-system calls calculate_drtm_pcr for legacy
  installations, using /boot/system/grub/grub.cfg and setting SRTM to 0.

This leads to PCR18 and PCR19 not being predicted correctly, but
predicted nonetheless while no error is apparently raised when
performing the upgrade/forward-seal.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>