tboot: handle pcr-calc tboot versions by ranges
The pcr-calc tool can be asked to emulate some changes that happened in
increasing versions of TBoot to predict the PCRs when upgrading to a new
version of TBoot.
This is better handled using ranges of versions to reduce the amount of
changes to the sources when upgrading TBoot. Refactor the version
handling logic to actually parse, compare and handle a numerical
"major.minor.micro" version number and take the appropriate actions
depending on the version.

Note: This is only relevant for TPM 2.0 upgrade from a TBoot 1.9.6 or
priori to a more recent TBoot.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>