xenmgr: regression on cmdline argument autostart
Xenmgr can be configured to not autostart VMs at startup, this used to
be handled by the initscript by checking "autostart" on the kernel
cmdline.
Commit: ae940b14 xenmgr: cleanup initscript
introduced a regression on a knee-jerk change from "autostart" to
"no-autostart" being checked on the command line, yet the configuration
files used at bootstrap were not changed.

This is in any case a regression that would affect any customised
openxt.cfg or grub.cfg, making it not worth it.

Restore the original behavior: normal operation boot entries add
"autostart" to the kernel cmdline, "no-autostart" is the default.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>