initscripts: fix dbus wait helpers
Commit:
2ef30c8c initscripts: dbus wait for service providers

Did not handle its purpose correctly.
dbus-send will not return an error code regardless of the method_call
result as long as the method was called successfully. Since the
signature returns a boolean this is easy to grep.

The timeout is expected to be roughly in seconds, but the loop does not
wait, so add that.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>