linux: syncvm: enable PVH mode
Enable XEN_PVH_GUEST so the syncvm can be run as virt-type PVH to avoid
spectre and meltdown issues.

To enable PVH, we also have to enable ACPI.  That brings in a bunch of
unnecessary ACPI defaults which are then turned off.

Dom0 support is also turned off since this isn't dom0.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>