xen: build freestanding to avoid multilib
The Xen recipes in meta-virt changed to requiring multilib in order to
compile hvmloader.  OpenXT picked this up when moving to OE Dunfell.

Meanwhile, upstream Xen 4.15 was patched to allow building hvmloader
without requiring 32bit libraries.

Backport the Xen patch and disable some of the meta-virt variables so
multilib is no longer required.  This avoids building lots of
(large) dependencies for the sole purpose of building hvmloader.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>