linux: blktap2: introduce blktap_destroy_work_common
blktap_ring_release and blktap_destroy_work have a common body.  It is
simple now, but will be expanded in a subsequent comment.  Put it into
blktap_destroy_work_common.

To do this, we need an initialized delayed_work, so move initalization
into blktap_ring_open.  With that, we have to move blktap_destroy_work
earlier to it is defined for that use.  blktap_ring_release and the new
blktap_destroy_work_common are also moved so they are all together.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>