monit: Run only in runlevel 5 for dom0
We want monit to run in dom0 to monitor vglass, disman and ivcdaemon.
Those components only run in runlevel 5.  If we don't do something,
monit would start them in console mode (runlevel 3).

Solve this by only running monit in runlevel 5.  It works for now since
monit is only watching vglass and friends.  The benefit of this is to
streamline monit's checks.  Monit doesn't have an internal notion of the
current runlevel, so it would have to check the current runlevel on each
interval which adds up.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>