monit: Add dom0-cfg to configure
Configure dom0 via a drop-in to /etc/monit.d to avoid having to import and
overwrite monitrc.

30 seconds is a little too long to wait for monit to discover the UI has
died.  Decrease it to 2 seconds.

Also use a delay of 12 seconds to give disman and vglass time to start
before monit attempts to start them.

Point monit paths to the volatile path /var/lib/monit, so it doesn't try
to use the default of /root/.monit

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>