layer.conf: Allow meta-selinux initscript bbappend
Undo the masking of meta-selinux initscript bbappend.

It was originally added in commit b318c939d52f because checkroot.sh
didn't exist.  But now the file does, so that is not a problem.

commit c43cb3a613c0 "layer: Amend comment on BBMASK in layer.conf."
expanded the comment to mention populate-volatiles.sh.  The
populate-volatiles.sh part we actually want so that /var/volatile is
labeled properly.  `touch /var/log/lastlog` is superfluous with the
volatiles creating it, so that is fine.

restorecon against the read-write /var/volatile and /run are fine.
/var/lib is read only, but labels were set already during image
creation.  Since the label is correct, restorecon won't try to modify
it.  /etc/resolv.conf is a readonly symlink to volatile, so that is also
okay.  /etc/adjtime does not exist, but restorecon `-i` will skip
non-existant files, so that is okay to leave.

We want this so that monit will have properly labeled /var/lib/monit ->
/var/volatile/monit access.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>