refpolicy: Enable and patch ifplugd module
We'll be using ifplugd, so enable the module.

We have a few avs denials with ifplugd.  ifup runs in ifplugd_t context
and is the busybox implementation.  It needs the same /run/ifstate.new
labeling as was done in commit 600242636b4d "refpolicy: Fix /run/ifstate
labeling"

ifplugs tries modprobing for the network device.  It will be autoloaded
when a vif is detected, so it is unnecessary.  Just dontaudit it.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>