initramfs-stubdomain: Update for Xen 4.14
Use vchan-socket-proxy for QMP access - this was taken from how Qubes
does their QMP access.  It's also the way that is built into Xen.

Read dm-argv for the arguments - that's how Xen 4.14 writes them.

Set the RDEPENDS to include the programs launched by init.  This pulls
them in as dependencies automatically.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>