xen: Update libxl-domain-state.patc for 4.14
An additional "running" change is needed for libxl_domain_unpause()
since the old was was renamed to libxl__domain_unpause_deprecated().
Without this, the VM stays in "creating" even when fully up.  For NDVM,
this means network configuration doesn't happen.

While at it, move the state update to after the unpause operation
succeeds.

The "suspended" one is also relocated for code changes.

Some minor fixups for s/ctx/CTX/ are also needed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>