xen: Use upstream stubdomain_cmdline patch
Xen 4.14 already had a stubdom_memory patch as stubdomain_memkb, so that
piece of libxl-stubdom-options.patch is no longer needed with xenmgr
changes.  stubdom_memory appended, and stubdomain_memkb sets, so this is
a behavior change.

Upstream xen also has a stubdomain_cmdline patch.  Backport and use
that instead of the stubdom_cmdline one.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>