libxl: Fixup `xl cd-insert` for stubdom
This is a replacement for the old libxl-iso-hotswap patch.  It is libxl
AO friendly, but it still relies on other OpenXT patches to only
attached the ISO to the stubdom.

This adds the steps of detaching the block device from stubdom and then
attaching the new (xvdc) device to the stubdom.  Some other domid tweaks
are needed since we are dealing with the stubdom and not the guest.

We also don't have json, so we have to swallow those EMPTY_JSON errors.

This relies on the stubdom adding /dev/xvdX to the assigned fdset
/dev/fdset/800Y.  That is done with busybox mdev and a script.  It's
done this way since Qubes already had some code for it, and it allows
seccomp to restrict open calls for QEMU.  i.e. Just passing /dev/xvdc
may work instead.  However, libxl was already doing the fdset thing, so
this keeps them more in synch.

The remove-fd call is needed to ensure that FD doesn't hold /dev/xvdc
open and prevent the disk detach for changes after the first.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>