Add stubdom side of iso-hotswap support
This wires up the support for adding detected /dev/xvd[a-d] to the respective
QEMU fdsets.  Those will be used by libxl QMP messages to change the
cdrom medium.

The hotswap process is:
QMP eject drive
xl block-detach PV disk in stubdom /dev/xvdc
xl block-attach PV disk in stubdom /dev/xvdc
QMP change drive to /dev/xvdc

The QMP command from libxl uses an fdset, so need QEMU in the stubdom to
have the expected fdset.

We use busybox mdev to run qemu-xvdc-add-fd.sh on add or remove.

On add, we open /dev/xvd* and pass the fd into QEMU over QMP via add-fd.

The fdsets are mapped from 8000-8003 with xvda/hda->8000, xvdb/hdb->8001
and so on.  So libxl know what to use.

For removal, libxl issues remove-fd so the FD is closed before the PV
remove step.  Otherwise the PV drive is held open, normal removal fails
and it falls back to force removal.  This throws things off.  The mdev
script for removal is after the kernel-level removal, which is too late
to close the FD.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>