initramfs-stubdomain: Fix stubdom networking
Xen 4.14 changes resulted in the stubdom bridge not getting set up.

Switch the implementation to be more in line with the Qubes approach
from https://github.com/QubesOS/qubes-vmm-xen-stubdom-linux/blob/1c88f84c50a5901cd784f85f22c5e9748957daa5/rootfs/init#L117
This will allow for seccomp in the future.

The approach is modified to handle multiple VIFs.  Instead of running
/etc/qemu-ifup, just put them into shell functions.

The MAC address generation is to to avoid duplicate MACs being seen in
NDVM.  Nothing should be sending from those interfaces, but this is
safer.

The MAC address lookup from OXT-275 is retained.  That was from when
xenvm was used, and xl doesn't seem to have a problem after xec-vm -n
$name --nic 0 delete.  xenmgr passes vifs without indices, and xl auto
assigns those.  However, linux doesn't necessarily assign eth0 to devid
0 - that seems to be an implementation quirk.  It is more robust to
cross reference the xenstore devid's mac to the linux interface, so
retain that.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>