xen: Remove allow-mwait-cstate.patch
allow-mwait-cstate.patch seems to be bogus.  It exposes MONITOR/MWAIT to
PV guests, but:

    * MONITOR never leaked into PV guests, as PV guests cannot
    * use the MONITOR/MWAIT instructions.  As such, they require
    * the feature to not being present in emulated CPUID.

    * Modern PVOPS Linux try to be cunning and use native CPUID
    * to see if the hardware actually supports MONITOR, and by
    * extension, deep C states.

MWAIT is only valid from Ring 0, but PV guest kernels run in Ring 1
(32bit) or 3 (64bit).  Hence "cannot use".  The deep C states only
matter for dom0 which is allowed without the patch.

Given the limited use of PV anyway, this isn't typically exercised
anyway.

At some point, some PV kernel must have used this, but we don't have a
patch in our tree to do something based on it. Any other distro is
upstream-based these days, and per the comment shouldn't be using this.
Dom0 is still allowed, and that is the one that can do some C-state
uploading.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>