linux: blktap2: Change destroy_work timeout
schedule_delayed_work takes a time in jiffies, so trying to divide HZ
leads to an incorrect timeout.

A 100ms timeout would actually be rather fast as it keeps triggering
while the device is open (bd_openers > 0).  Since this is just a delayed
cleanup operation, bump it to once every 5 seconds.  With blkback
holding the tapdev open, cleanup won't happen until the VM shuts down or
is destroyed.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>