xen: Restore libxl json usage
It's unclear what broken with it present.  libxl expects it, so allow
it instead of potentially breaking something inside.

Maybe we will have errors with SELinux?

There used to be Xenstore re-writing for NICs and atapi disk
hotplugging.  Those are both gone now, so maybe that won't throw off the
json?  It's unclear what exactly it was.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>