xen: libxl: Fix argo_xswait_cb
argo_xswait_cb isn't actually waiting.  The xswait machinery calls
argo_xswait_cb multiple times for the duration of the timeout.  The
first call is an empty string because the entry does not exist.  We
should just return in that case to continue waiting, whereas we just
continue on today.

The strcmp check needs to be inverted to not match "1".  If we get an
unexpected value, we return which may eventually time out if
it doesn't change.

When we do have an error, we should return rc so it directly by calling
qmp_proxy_spawn_outcome instead of starting more work to then error out.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>