seal-system: Mark p local in pcr_in_selection
p is used as a local iterator in pcr_in_selection.  However, it was not
marked as local, so it can clobber uses of p in calling functions like
the forward seal code.

Mark it local.  This will be useful when using pcr_in_selection within
a loop in the forward seal code.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>