init.root-ro: Provide info when Unseal fails
When measured launch fails, users are in the dark as to what failed.
Try to help them out by providing info into which PCRs failed the
unseal.

Users shouldn't trust the compromised computer, so they shouldn't trust
this output.

This is a bad idea or terrible idea?  It is useful for debugging since
otherwise you need to do more work.  Also, it avoids needing to
calculate the capped PCR-15 value after unlocking.

This is based off the unencrypted forward_pcr.lst, so it is based off
untrusted input.  It would be "better" to parse the sealing set out of
the sealed data blob.  That is also just lying on the hard drive
unencrypted.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>