xen: libxl-syslog print errno
libxl macros pass errno to the logging functions.  It gets passed
through xtl_log and will be printed on the end of the line by the stdio
logger.  The OpenXT syslog implementation doesn't call xtl_log and is
just dropping it.

Add a comparable implementation to use strerror_r and add it to the
message passed to syslog.  libxl is using the GNU specific strerror_r,
which may return a read-only string and not update the passed-in buffer.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>