xen: libxl: Use garbage collection for pci_dev_wrap
xl was observed to abort with "malloc(): unsorted double linked list
corrupted".  It was down in xlu_cfg_readdata after rebooting a domain.

This may or may not be the cause.

pcidevs is added to the gc, but it is also manually freed.  This could
have corrupted things, though why it wasn't caught on double free is
unclear.  Regardless, it should only be free-ed once.

Between 4.14 and 4.16, libxl_device_pci_list_free was introduced.  Use
that to free the pci_list.

While doing this, use libxl_malloc to allocate pciw itself and let
garbage collection handle its free-ing.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>