xen: libxl: Move PCI reset into destroy_finish_check
Devices passed to a stubdom aren't getting reset.  For an HVM & stubdom,
the PV PCI front-back entries are to the PV stubdom, and QEMU in the
stubdom provides the PCI emulation into the HVM.

This means PCI devices show up under stubdom's xenstore entries.  This
means the pci wrap is filled in for stubdom but not the main domain.
The current code only tries to reset PCI devices for the domain and not
the stubdom.

Move the reset code into a function and call it from
destroy_finish_check for both the domain and stubdomain.  This will
cover domains with and without stubdoms.

There was a problem with the NULL check being omitted reset_pci_devices,
and then libxl would segfault on a NULL pointer, right before a ud2
intruction.  It was strange, but making the NULL checks in
destroy_finish_check worked.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>