xen: Drop prune-vga-acpi-dev.patch
This patch's description is at odds with the contents.  The description
says it is to fix S3 on Win2k3/XP, but the removed code says it supports
S3 on those same OSes.

With or without this patch, Windows 10 `powercfg /a` reports:
Standby (S3)
    An internal system component has disabled this standby state.
             Graphics

Trying to have the _S3D method return S3, like QEMU does when VGA is
QXL, did not change the above.  This may just be a limitation of the
VGAPNP.sys driver.

Just drop this patch.  It doesn't help current supported Windows
versions, so lets remove some cruft from our patch queue.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>