qemu: Pass CFLAGS for security_flags
QEMU's configure fails when security_flags are enabled because fortify
source is specified, but no optimization is turned on (which is a
wanring that fails with -Werror).  Bitbake specifies from flags in CC
and others in CFLAGS.  QEMU configure uses CC, but not CFLAGS, which
causes the problem.  The proper way to apss CFLAGS is to use
--extra-cflags.  Use that, which is what OE does for its qemu recipe and
allows configure to complete.

This also means the CFLAGS_append wasn't taking affect, AFAICT, so drop
it.  security_flags is applying hardening that is _append was trying to
do.

With this, drop -DXC_WANT_COMPAT_DEVICEMODEL_API as it is no longer
needed.  QEMU can compile without it since QEMU 2.10 (with Xen 4.9 and
later).

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>