qemu/xen: Remove ACPI SX customization
The ACPI \_S4 method defines the values that the OS should write to
PM1a/b_CNT to enter S4.  The same for S3 and S4

QEMU already has some support - it uses 1 for S3 and 0 for S5.  S4 is
configurable via s4_val, but it defaults to 2.

All the S3/4/5 customization in qemu-dm/acpi.patch and
hvm-pm-hibernate-s-state.patch is unnecessary.  QEMU still needs to
write the xenstore_power_state, but we can just use the existing code
there.  The Xen ACPI code just needs to patch in value 2 instead of 6
and the VMs can S4 with less customization.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>