linux: Drop netfront-support-backend-relocate.patch
Like xenbus-move-otherend-watches-on-relocate.patch,
netfront-support-backend-relocate.patch was written when xenvm was used.
xenvm had some code to directly re-write vif xenstore entries to move
the backend.  Since the xl conversion, vifs are detached and re-attached
to move, so this code is no longer needed.  Since it is an OpenXT patch,
it wouldn't help actual guest VMs.  And the patch description's
reference to UIVM points to it no longer being relevant to our vif-less
UIVM.

Remove the patch.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>