linux: Drop xenkbd-tablet-resolution.patch
xenkbd-tablet-resolution.patch changes the default vkb pointer
resolution from 800x600 to 32767x32767.  As of xenmgr manager.git commit
1bdfaddf22d7 "xenmgr: Set feature-abs-pointer for vkb", the xenstore
entries for height and width are set for 32767x32767.  It's outside of
this patches context, but those are read into ptr_size[KPARAM_X] (and
Y).  So dropping this patch will still pass 32767 to
input_set_abs_params(), and we can drop it.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>