qemu-dm-wrapper: Fix QEMU pid
There is a problem with the wrapper.  libxl tracks device-model-pid to
later kill the QEMU process if necessary, but 'exec qemu | logger' isn't
doing what we want.  Since the qemu-dm-wrapper shell script is using a
pipe/pipeline, it is actually forking qemu and logger each in their own
subshell.  `exec qemu` then runs in that subshell, but the parent
qemu-dm-wrapper retains the PID libxl is tracking.  libxl sends sighup
to qemu-dm-wrapper which exits, but QEMU keeps running.

With the use of xen scripts to run block-tap, this causes a lockup.
The block-tap script grabs the block lock, and calls tap-ctl destroy.
tap-ctl hangs because QEMU still holds a reference to
/dev/xen/blktap-2/tapdevN.  tap-ctl has an inheritted FD open to the
block lock which prevents any other process from grabbing the lock.

To work around libxl signalling qemu-dm-wrapper and not qemu itself,
just catch the SIGHUP and forward to the child QEMU process.

Fixes: bbbb4b6296bb "qemu-dm: Drop syslog patch"

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>