xen: xl: Use /dev/null for stdout/stderr when daemonized
xenmgr launches `xl create` which daemonizes to track the life
of the domain.  The parent `xl create` was launched with stderr attached
to a pipe.  When the parent exits, xenmgr closes the pipe.

Today when xl calls do_daemonize, it only re-directs stdin to /dev/null.
When the daemonized xl detects the shutdown, it tries to write to
stderr.  When a stronger SELinux policy in enforcing mode, this write
generates SIGPIPE which cause xl to exit and the domain is not cleaned
up.

Re-assigning stdout & stderr when daemonizing avoids this and is
matches the un-patched behaviour of dup2()ing logfile to those FDs.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>