xen: Have xl ignore SIGPIPE
XenMgr runs xl commands to control domains.  Today everything works.
However, future changes can have xl write to stderr, which is a dangling
pipe.  This generates SIGPIPE and terminates the xl process.

The failure was noted when `xl shutdown` is run via
runInteractiveCommand_closeFds to avoid leaking FDs, and the SELinux
policy is tigthened up and enforcing.  Strangely not-enforcing doesn't
have terminal SIGPIPEs signals.

Logging is through syslog, so any writes to stderr are superfluous.
Ignore SIGPIPE so they are non-fatal.  The write is not removed since
an administrator running xl in a terminal would like to see the output.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>