dom0-tweaks: Add delayed-seal initscript
delayed-seal is an initscript to run on shutdown after VMs have been
stopped.  We don't want dom0 doing MMIO operations with the TPM while a
guest is running that could potentially snoop on the data.

It has additional code to try destroying any still-running VMs as a
safety measure.

Needs the installer to touch /tmp/upgrade-delayed-seal to trigger the
sealing.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>