Control PCR debugging with debug-tweaks
Include the code to print a diff of the PCRs during measured launch
failure, but have it hardcoded off.  When debug-tweaks is set in
IMAGE_FEATURES, the xenclient-dom0-image recipe will set the flag and
enable the debugging feature.

This can be activated by setting
EXTRA_IMAGE_FEATURES_pn-xenclient-dom0-image = " \
    debug-tweaks \
"
in conf/local.conf

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>