xen: Add vchan reconnect marker patches
A usbvm was failing to start on a machine with two USB controller PCI
devices.  It wasn't an issue with the stubdom missing the Connected
xenstore state.

Marek wrote these vchan patches, so pull them in.  The error messages
match, though there isn't much detail.  Testing ~60 starts of the usbvm
have succeeded, so this seems to have addressed the issue.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>