Default to read-only blktap
Change the default for `tap-ctl create` to read-only, the former -R
option.  Add a new -W option to use when read-write is desired.

Now that tap-ctl defaults to read-only, we need to specify -W when
opening writable .vhds.  For the read-only cases, explicity add -R to
highlight the use.

OpenXT is changing to hash .vhds directly.  To do so, we need to only
open .vhds as read-only as read-write can modify the metadata even if
the contents do not change.  To avoid user error, default to read-only
so explicit action is needed which will potentially modify the .vhd.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>