[ghc8] xenmgr: Point to the correct templates directory
xenmgr uses the getDataFileName function to get the /usr/share
dir that the templates reside in. This function is autogenerated
at buildtime according to the $datadir and $datasubdir variables.

$datadir defaults to /usr/share, which is correct, and $datasubdir
defaults to x86_64-linux-ghc-8.10.7, which is not. Set this variable
to xenmgr-1.0 instead to follow the expected format defined by the
recipe.

Signed-off-by: Nicholas Tsirakis <tsirakisn@ainfosec.com>