grub: Recipes for grub xen pvh and pv64
Based off the Qubes configuration, but slightly stripped down.  Qubes
bundles in many more modules.

This will let us avoid parsing the guest disks to get the kernel and
instead let grub do that inside the VM.

The weird FILESEXTRAPATHS is to allow the recipes in this layer to find
make the SRC_URI-specified files in the OE-core layer.

grub2 doesn't build for pv32, so it is not included.