Add grub-xen-conf for VMs
This introduces a new recipe grub-xen-conf that creates
/boot/grub2/grub.cfg.  This is used by grub-xen-pvh or grub-xen-pv64 to
load the VMs.

A default APPEND is supplied that works for the non-partitioned uivm and
syncvm.  If a machine specifies one, it should be picked up.

This allows changing syncvm & uivm to be booted by grub-xen-pvh instead
of extracting their kernels.  ndvm can be booted by grub-xen-pv64 in pv
mode.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>