linux: dom0: Add input drivers
Some HP laptops have a message about enabling RMI4 and Synaptic for
better touchpad support, so do so.

Some Dells have an i2c-connected touchpad that is "better" than just the
psmouse, so enable more drivers for that - hid-multitouch &
i2c_hid_acpi.

More input stuff is enabled since it seemed necessary and then also
others may be encountered in the future.

Note that intel-hid and intel-vbtn are not enabled.  They can generate
KEY_POWER for the power button.  However, instead of xcpmd seeing that,
vglass sees it and passed it to the guest.  So you can shutdown just
uivm and not the platform.  Just don't include the driver.