Ensure directories/mountpoints exist for initial scripts
With changes to crypsetup (LUKS2), the boot timing seems to be altered slightly - /run now needs to be mounted earlier to allow cryptsetup calls in unlock_config to pass, and /storage needs to be mounted earlier so that xenmgr.initscript succeeds in creating null.iso. Also add safeguard to ensure they exist when needed.